Add overcommit.yaml (#41)
Since we're going to be doing code review through GitHub instead of
Phabricator for this project, we're going to leave behind some of the
tooling that we've built into arc/phab.

Overcommit is a generic git hook framework I helped build/open source at
a previous company, and is pretty awesome in general. This will allow us
to run go vet / go lint during commits, as well as add custom hooks.
I've also enabled YAML and JSON validation as an example of other things
we can do with it.