make: Always use vendored thriftrw and plugins (#133)
* make: Always use vendored thriftrw and plugins

This changes `make examples` to always use the vendored `thriftrw` and
`thriftrw-yarpc-plugin`. It took me a while to realize that `make test` was
failing to build because it was using my globally installed version of
`thriftrw` and `thriftrw-yarpc-plugin`, both of which are newer versions than
what UberFx is using.

The `vendor-build.sh` script added here is also usd in yarpc-go to do
something similar. It is a simplified version of an internal tool which we are
planning on open sourcing in the future.

* vendor-build: Work without realpath

The realpath command isn't available on Travis

* rename to build_vendored.sh

* build_vendored docs

* Fix glide install