Restore logging functionality (#341)
* Restore dev logging functionality

A couple of things was going on here after the large zap refactor. The
most important one is the populateStruct error was not the only
condition to use the default logger, the other is on logging
configuration at all.

Logging setup now returns an error if it's not able to properly parse
the config.

* Fix the broken test and zap serialization

At it's core, what was happening is that zap.Config was an embedded
object of LogginConfiguration. Embedding objects doesn't result in the
desired yaml serialization, as internal ones are just not initialized at
all and all the attributes were being discareded

* Use zap defaults for log config

Re-enable and test Sentry DNS serialization

* Change default encoding to json

* Fix error message [ci skip]

* Production config already outputs json [ci skip]

* Add a couple more keys to logging config test

* Move logging config into ulog package

* Fix lint

* Add TODO with ticket number [ci skip]