Expand environment variables only in static yaml files (#486)
YAML provider replaces all the strings in form ${var} $var everywhere, even in comments and there is no way to opt out of it.

Instead of a blank substitution we introduce separate constructors for interpolation. WithExpand constructors will iterate over all the yaml nodes and replace only values, that need to be interpolated.

Loader is updated to reflect this changes, it has a staticFiles field with a corresponding Set method to set files that will not be interpolated, remaining base file will be expanded with the os.Lookup function.