Add fx.Inject to pull values from the container (#504)
This adds an `fx.Inject(target)` function, the output of which may be
passed to `App.Start()` to fill the target struct with values from the
DI container.

    var x struct {
        Dispatcher *yarpc.Dispatcher
    }

    app.Start(ctx, Inject(&x))

This enables something very simple that users will expect to be able to
do: start the application, do things with the objects inside the started
application, and then stop it--all in a serial manner rather than
out-of-band in a separate goroutine. Need for this has already become
apparent while writing integration tests that interact with Fx.
See #500 for more discussion.

This works by generating a function that accepts a single argument for
each exported field in the struct. The generated function copies values
from the received arguments back to the user-provided struct, leaving
unexported fields alone.