Run Invokes during New, not during Start (#551)
This builds on top of a lot of previous work, most notably the refactor
of the top level API to no longer expose `Provide` and `Invoke` on the
app object. Now that all options have to atomically be provided on
`New`, it no longer makes sense to wait until the application start to
resolve.

All the heavy-weight work, like listening for the network, will still be
done as part of the `OnStart` hooks.

Fixes #550.

Essentially this introduces a new touch point: when the graph is fully
resolved, but the start Hooks have not yet run.