Drop closure-based fx.Option implementation (#686)
We have a handful of `fx.Option` implementations based on `optionFunc`
which captures value using a closure. That makes them opaque and
uninspectable.

This change drops the `optionFunc` type in favor of concrete
implementations for all `fx.Option`s. For further inspection, each such
instance includes a `String()` implementation, adding `fmt.Stringer` to
the `fx.Option` interface in the process.