fxreflect: Delete ReturnTypes in favor of dig.Info (#750)
With uber-go/dig#280 done, we can use the ProvideInfo produced by Dig to
determine what types will be produced by a constructor and log them, so we
no longer need to reimplement the `dig.Out` inspection logic in Fx.

To surface this information with the new event API introduced in #747, this
adds a new field to fxevent.Provide:

    OutputTypeNames []string

This, in lieu of exposing `dig.Info` wholesale ensures that we have some
wiggle room here.

Note that this introduces one behavioral difference on when the Provide
event is logged: previosly, the Provide event would be logged whehter it
succeeded or failed.

With this change, it will be logged only if it succeeds -- after it
succeeds. This makes sense because otherwise we can't extract information
about the function from Dig.

This temporarily pins Fx to the unreleased `master` branch of Dig. We'll
have to tag a releaes of Dig before we can release this feature.

Refs GO-676